Update patch set 2
Patch Set 2:

Hi Guillem, could you please review the package and apply necessary adjustments, if any. The package itself it is quite lightweight but I have not tested it yet, sorry for that. Alex will merge it once it is ready. Very much appretiate it! :) If you have questions regarding the package please ask Alex as I am on vacation since today. Thanks alot to both of you guys! :)

Patch-set: 2