Update patch set 1
Patch Set 1:

localhost/127.0.0.1 is not be used for that in Carrier as ha_int, but yes auto skipping it for localhost/127.0.0.1 for now solves it for PRO (but not Carrier).
an idea: we can add another option there then
"ssl_verify_fqdn_only"
so if someone wants to use an IP address for that (thats what all our carriers this is the case afaik) they just enable "ssl_verify_fqdn_only".
I am not sure that for Carrier if someone wants to provide an IP address - their domain cert will work as it is likely that for the domain they have a different IP/range than what "ha_int" is assigned to.

Patch-set: 1
Label: Code-Review=0