TT#53701 Transition perl module package into a libfoo-perl package
The current ngcp-api-tools package name seems to indicate this contains
some actual tool, but it only provides perl modules. Let's rename it to
clarify its actual intent.

At the same time we demote the ngcp-panel from a Depends to a
Recommends, because this is an HTTP client and can be used from a host
that has no panel installed against a host that has it. And remove the
ngcp-api-tools provides as we are not providing any tool.

Change-Id: I3fd3854a91cc680f71d5465bc7c397f292f7b95a