Update patch set 2
Patch Set 2:

this conversation generation files should not be part of the package.
they will be as a part of ngcp-dev-tools.
as I wrote in the other review: no need from your side to do anything extra there as some parts of it are now on me, we get back to it in next sprints.
for this review only the Client.pm upload implementation should be there, please merge Client.pm only after review and an approval/changes by Gerhard.

Patch-set: 2