Update patch set 9
Patch Set 9:

we are not going to commit ngcp-conversation related into ngcp-api-tool, it will go into ngcp-dev-tools as it has nothing to do with using on the production customer platforms but just for internal developent.
but lets take care of that when I do the ngcp-dev-tools related redesign of that (hopefully next or in a sprint).

Patch-set: 9