TT#68300 Load the configuration file from the constructor
We should not be running code that loads a configuration file in
a BEGIN block, as that will execute just by loading the module.

This makes using this module conditionally in other programs or testing
that code more difficult, as we would need to mock either this module
or the configuration file.

Just load the configuration file once in the constructor and cache
its contents, to avoid the overhead.

Change-Id: I86986dc81eaa32833b738ffa6df2a5de30159482