Update patch set 2
Patch Set 2:

> LGTM, switching from JSON to JSON:XS might brake backward
 > compatibility here, isn't it?

The JSON module is mostly a transparent wrapper over JSON::XS or JSON::PP (the pure perl implementation). And it defaults already to use JSON::XS if it is available. This just skips the intermediary, and uses the non-deprecated functions.

 > Maybe we should announce it as technical info?

I don't think there's any need. If there's any problem here that would be due to the difference in encoding handling from the deprecated functions, which should be fixed here, but I think the conversion is fine.

Patch-set: 2
Reviewer: Gerrit User 1000016 <1000016@85315e57-5f03-49c3-83f8-201ae2313a75>