TT#68300 Fix programs --help output, do not print missing help parameter
The GetOptions() function will pass «$opt_name, $opt_value» to the
function refs in the options spec, which means that the usage()
function will consider that a missing parameters message and print
"help" as being a missing required parameter.

Use an anonymous function to wrap the usage() call and avoid the
unexpected arguments.

Change-Id: I5c10b241da23ccb7aa25a13dd5ebc1b4188b7370