Update patch set 1
Patch Set 1:

> keep an eye on this, this part was topic for the disaster of
 > http://gerrit.mgm.sipwise.com/c/ngcp-api-tools/+/34849

I've checked the other link you gave, and I'm not sure I see the relation here TBH. The only think that might be worth considering is whether on an HTTP request error, we want to reset the iterator (the undef just before the new code), or not, in case we want to allow the call site to retry or just cancel the current iteration. I went for the latter, but I can see the former being also a valid approach.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@85315e57-5f03-49c3-83f8-201ae2313a75>